refactor: 전반적인 리팩터링
